+++ packages/core/src/serialization/codecs/GenericChangeCodec.ts
@@ -16,15 +16,14 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import ObjectCodec from '../../serialization/ObjectCodec';
+import ObjectCodec from '../ObjectCodec';
 import { isNode } from '../../util/domUtils';
-import Codec from '../../serialization/Codec';
+import type Codec from '../Codec';
 
 /**
- * Codec for {@link ValueChange}s, {@link StyleChange}s, {@link GeometryChange}s,
- * {@link CollapseChange}s and {@link VisibleChange}s. This class is created
- * and registered dynamically at load time and used implicitly
- * via <Codec> and the <CodecRegistry>.
+ * Codec for {@link ValueChange}s, {@link StyleChange}s, {@link GeometryChange}s, {@link CollapseChange}s and {@link VisibleChange}s.
+ *
+ * This class is created and registered dynamically at load time and used implicitly via {@link Codec} and the {@link CodecRegistry}.
  *
  * Transient Fields:
  *
@@ -34,16 +33,13 @@ import Codec from '../../serialization/Codec';
  * Reference Fields:
  *
  * - cell
- *
- * Constructor: GenericChangeCodec
- *
- * Factory function that creates a <ObjectCodec> for
- * the specified change and fieldname.
- *
- * @param obj An instance of the change object.
- * @param variable The fieldname for the change data.
  */
-class GenericChangeCodec extends ObjectCodec {
+export class GenericChangeCodec extends ObjectCodec {
+  /**
+   *
+   * @param obj An instance of the change object.
+   * @param variable The field name for the change data.
+   */
   constructor(obj: any, variable: string) {
     super(obj, ['model', 'previous'], ['cell']);
     this.variable = variable;
@@ -54,7 +50,7 @@ class GenericChangeCodec extends ObjectCodec {
   /**
    * Restores the state by assigning the previous value.
    */
-  afterDecode(dec: Codec, node: Element, obj: any): any {
+  afterDecode(dec: Codec, _node: Element, obj: any): any {
     // Allows forward references in sessions. This is a workaround
     // for the sequence of edits in mxGraph.moveCells and cellsAdded.
     if (isNode(obj.cell)) {
@@ -65,5 +61,3 @@ class GenericChangeCodec extends ObjectCodec {
     return obj;
   }
 }
-
-export default GenericChangeCodec;
