+++ packages/core/src/view/GraphDataModel.ts
@@ -32,8 +32,6 @@ import TerminalChange from './undoable_changes/TerminalChange';
 import ValueChange from './undoable_changes/ValueChange';
 import VisibleChange from './undoable_changes/VisibleChange';
 import Geometry from './geometry/Geometry';
-import ObjectCodec from '../serialization/ObjectCodec';
-import CodecRegistry from '../serialization/CodecRegistry';
 import { cloneCells, filterCells } from '../util/cellArrayUtils';
 
 import type { CellStyle, FilterFunction } from '../types';
@@ -1200,61 +1198,4 @@ export class GraphDataModel extends EventSource {
   }
 }
 
-/**
- * Codec for <Transactions>s. This class is created and registered
- * dynamically at load time and used implicitly via <Codec>
- * and the <CodecRegistry>.
- */
-export class ModelCodec extends ObjectCodec {
-  constructor() {
-    super(new GraphDataModel());
-  }
-
-  /**
-   * Encodes the given <Transactions> by writing a (flat) XML sequence of
-   * cell nodes as produced by the <CellCodec>. The sequence is
-   * wrapped-up in a node with the name root.
-   */
-  encodeObject(enc: any, obj: Cell, node: Element) {
-    const rootNode = enc.document.createElement('root');
-    enc.encodeCell(obj.getRoot(), rootNode);
-    node.appendChild(rootNode);
-  }
-
-  /**
-   * Overrides decode child to handle special child nodes.
-   */
-  decodeChild(dec: any, child: Element, obj: Cell | GraphDataModel) {
-    if (child.nodeName === 'root') {
-      this.decodeRoot(dec, child, <GraphDataModel>obj);
-    } else {
-      this.decodeChild.apply(this, [dec, child, obj]);
-    }
-  }
-
-  /**
-   * Reads the cells into the graph model. All cells
-   * are children of the root element in the node.
-   */
-  decodeRoot(dec: any, root: Element, model: GraphDataModel) {
-    let rootCell = null;
-    let tmp = root.firstChild;
-
-    while (tmp != null) {
-      const cell = dec.decodeCell(tmp);
-
-      if (cell != null && cell.getParent() == null) {
-        rootCell = cell;
-      }
-      tmp = tmp.nextSibling;
-    }
-
-    // Sets the root on the model if one has been decoded
-    if (rootCell != null) {
-      model.setRoot(rootCell);
-    }
-  }
-}
-
-CodecRegistry.register(new ModelCodec());
 export default GraphDataModel;
