+++ packages/core/src/view/undoable_changes/CollapseChange.ts
@@ -16,8 +16,6 @@ limitations under the License.
 
 import Cell from '../cell/Cell';
 import GraphDataModel from '../GraphDataModel';
-import CodecRegistry from '../../serialization/CodecRegistry';
-import GenericChangeCodec from './GenericChangeCodec';
 
 import type { UndoableChange } from '../../types';
 
@@ -52,8 +50,4 @@ class CollapseChange implements UndoableChange {
   }
 }
 
-const __dummy: any = undefined;
-CodecRegistry.register(
-  new GenericChangeCodec(new CollapseChange(__dummy, __dummy, __dummy), 'collapsed')
-);
 export default CollapseChange;
